Proof-of-Concept support for Twitch command through Helix

+++ Actions/Dashboard/IActionUser.cs
@@ -7,6 +7,6 @@ namespace Actions.Dashboard
         string ID { get; }
         string Name { get; }
         string? ProfilePictureURL { get; }
-        Task Ban(float? length);
+        Task Ban(uint? length);
     }
 }
